+++ Engine/Runtime/src/Extra/SunPointShadowMapRenderingNode.cpp
@@ -4,38 +4,38 @@
 #include "SunPointShadowMapRenderingNode.h"
 
 SunPointShadowMapRenderingNode::SunPointShadowMapRenderingNode() {
-	
+
 }
 
 void SunPointShadowMapRenderingNode::init() {
 	shadowMapShader = SunShader("./Graphics/Shaders/Variable Pipeline/Scene.vert", "./Graphics/Shaders/Variable Pipeline/Scene.geom", "./Graphics/Shaders/Variable Pipeline/Scene.frag", "./PointShadowMap.pre");
-	
+
 	addAction("loadPointLights", &SunPointShadowMapRenderingNode::loadPointLights);
-	addAction("render", &SunPointShadowMapRenderingNode::render);	
+	addAction("render", &SunPointShadowMapRenderingNode::render);
 }
 
 void SunPointShadowMapRenderingNode::loadPointLights(SunAction action) {
 	// Action
-	SunAction sceneAction("initializeShadowMapRenderer");	
-	
+	SunAction sceneAction("initializeShadowMapRenderer");
+
 	// Renderer
-	sceneAction.addParameter("renderer", this); 
-	
+	sceneAction.addParameter("renderer", this);
+
 	//sendAction(sceneAction, scene);
 }
 
 void SunPointShadowMapRenderingNode::render(SunAction action) {
 	// Action
-	SunAction shadowMapAction("shadowMap");	
-	
+	SunAction shadowMapAction("shadowMap");
+
 	// Shader Map
 	map<string, SunShader> shaderMap = {{"scene_all", shadowMapShader}};
-	shadowMapAction.addParameter("shaderMap", &shaderMap); 
-	
+	shadowMapAction.addParameter("shaderMap", &shaderMap);
+
 	// Scene
-	//shadowMapAction.addParameter("scene", scene); 
-	
-	for (int i = 0; i < lights.size(); i++) {
+	//shadowMapAction.addParameter("scene", scene);
+
+	for (size_t i = 0; i < lights.size(); i++) {
 		sendAction(shadowMapAction, lights[i]);
 	}
 }
